move back to player crate

+++ libplatune/player/src/dto/player_state.rs
@@ -1,6 +1,6 @@
 #[derive(Clone, Debug)]
 pub struct PlayerState {
-    pub volume: f32,
+    pub volume: f64,
     pub queue: Vec<String>,
     pub queue_position: usize,
 }
